requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic~=1.10.7
passlib[bcrypt]==1.7.4
python-jose~=3.3.0
alembic==1.9.4
python-dotenv==0.21.0

starlette~=0.26.1
config~=0.5.1